+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Add `Program#attach_tcx`
 - Add `Program#attach_xdp`
 - Add `Program` and `Link` classes with `Program#attach`